Fixed compress_and_checksum

+++ compress_and_checksum.py
@@ -17,6 +17,7 @@ from locale import getpreferredencoding
 # Path stuff
 from os.path import join as os_join
 from os.path import dirname as os_dirname
+from os import remove as os_remove
 from os import environ
 from pathlib import Path, PurePath
 from inspect import getsourcefile
@@ -44,7 +45,8 @@ STAGING_DIR = os_join(SCRIPT_DIR, "Staging")
 CHECKSUMS_DIR = os_join(STAGING_DIR, "Checksums")
 COMPRESSED_DIR = os_join(SCRIPT_DIR, "Compressed")
 
-# TODO CPU/4 for max_workers
+# Set up our threadpool executor so it has NUM_CPUS/2 threads
+# because we use 7z with 2 threads
 available_cpus = len(sched_getaffinity(0))
 EXECUTOR = ThreadPoolExecutor(max_workers=ceil(available_cpus/2))
 
@@ -77,16 +79,21 @@ def sevenz_archive(entrypoint, files_and_or_dirs, destination_file):
         sevenz_exec = ["7z", "a", "-bd", "-bb0", "-mx=9", "-mmt=2"]
     appended_files_and_or_dirs = []
     for file in files_and_or_dirs:
+        print(file)
         appended_files_and_or_dirs.append(os_join(entrypoint, file))
     print(f"Compressing: {destination_file}")
-    subprocess.run(sevenz_exec + [destination_file, appended_files_and_or_dirs],
+    subprocess.run(sevenz_exec + [destination_file] + appended_files_and_or_dirs,
                    capture_output=False,
                    check=True)
 
     if DELETE_AFTER_COMPRESS:
         for target in appended_files_and_or_dirs:
             print(f"Removing: {target}")
-            rmtree(target)
+            my_file = Path(target)
+            if my_file.is_dir():
+                rmtree(target)
+            if my_file.is_file():
+                os_remove(target)
 
     calculate_and_write_checksum(destination_file)
     return
@@ -94,9 +101,9 @@ def sevenz_archive(entrypoint, files_and_or_dirs, destination_file):
 
 def main():
     rs_var_dump = load_rs_var_dump()
-    rs_liveries = rs_var_dump.rs_liveries
-    rsc_liveries = rs_var_dump.rsc_liveries
-    roughmets = rs_var_dump.roughmets
+    rs_liveries = rs_var_dump['rs_liveries']
+    rsc_liveries = rs_var_dump['rsc_liveries']
+    roughmets = rs_var_dump['roughmets']
 
     # Red Star BIN
     EXECUTOR.submit(sevenz_archive,
@@ -108,14 +115,14 @@ def main():
     for roughmet in roughmets:
         EXECUTOR.submit(sevenz_archive,
                         roughmet['roughmet_directory'],
-                        [file for file in roughmet['files']],
+                        roughmet['files'],
                         os_join(COMPRESSED_DIR, f'{roughmet["roughmet_directory_basename"]}.7z'))
 
     # Red Star Liveries (camo and black)
-    for livery in [rs_liveries + rsc_liveries]:
+    for livery in rs_liveries + rsc_liveries:
         EXECUTOR.submit(sevenz_archive,
                         os_join(STAGING_DIR, livery["dcs_airframe_codename"]),
-                        [[livery["livery_base_dirname"]] + livery["livery_pilot_dirs"]],
+                        [livery["livery_base_dirname"]] + livery["livery_pilot_dirs"],
                         os_join(COMPRESSED_DIR, f'{livery["livery_base_dirname"]}.7z'))
 
     EXECUTOR.shutdown(wait=True)
